% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{AnnotateHybrid}
\alias{AnnotateHybrid}
\title{Annotate hybrid data.table}
\usage{
AnnotateHybrid(hybrid.dt)
}
\arguments{
\item{hybrid.dt}{hybrid data.table to annotate}
}
\value{
annotated data.table
}
\description{
Annotates each arm of a hybrid data.table using rna.tc.gr
}
