% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{AnnotateHybridArm}
\alias{AnnotateHybridArm}
\title{Annotate GRanges arm}
\usage{
AnnotateHybridArm(arm.gr)
}
\arguments{
\item{arm.gr}{GRanges to annotate}
}
\value{
Annotated GRanges object
}
\description{
Annotates GRanges arm (created using ConvertToGRanges) using rna.tc.gr
}
