% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{ConvertToDataTable}
\alias{ConvertToDataTable}
\title{Convert GRanges to data.table}
\usage{
ConvertToDataTable(gr)
}
\arguments{
\item{gr}{GRanges to convert}
}
\value{
data.table or hybrid data.table
}
\description{
Converts from GRanges to data.table. If only one arm then returns single data.table, if both arms then returns hybrid data.table
}
