% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{ConvertToGenomicCoordinates}
\alias{ConvertToGenomicCoordinates}
\title{Convert to genomic coordinates}
\usage{
ConvertToGenomicCoordinates(hybrid.dt)
}
\arguments{
\item{hybrid.dt}{hybrid data.table}
}
\value{
GRangesList of length two (L and R arms)
}
\description{
Convert from transcriptomic to genomic coordinates
}

