% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{ConvertToGRanges}
\alias{ConvertToGRanges}
\title{Convert data.table to GRanges}
\usage{
ConvertToGRanges(hybrid.dt, arm = c("all", "left", "right"))
}
\arguments{
\item{hybrid.dt}{hybrid data.table to convert}

\item{arm}{arm to convert (all, left, right)}
}
\value{
GRanges object
}
\description{
Convert data.table to GRanges keeping metadata
}
