% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{CreateBED}
\alias{CreateBED}
\title{Create BED file}
\usage{
CreateBED(gc.grl, filename)
}
\arguments{
\item{gc.grl}{GRangesList of hybrids in genomic coordinates}

\item{filename}{Filename to save}
}
\value{
BED file
}
\description{
Export genomic coordinate BED file from converted coords output
}

