% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CreateDirectories}
\alias{CreateDirectories}
\title{Create directory structure}
\usage{
CreateDirectories(folder)
}
\arguments{
\item{folder}{Folder in which to create directory structure}
}
\value{
New directories
}
\description{
Create directory structure
}

