% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{CreateNonRedundantHybrids}
\alias{CreateNonRedundantHybrids}
\title{Create non-redundant set of hybrids}
\usage{
CreateNonRedundantHybrids(islands, nonislands)
}
\arguments{
\item{islands}{data.table of islands}

\item{nonislands}{data.table of non-island hybrid reads}
}
\value{
annotated data.table
}
\description{
Join island and non-island hybrids to create non-redundant set of hybrids
}
