% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes_mapping_functions.R
\name{DemultiplexReads}
\alias{DemultiplexReads}
\title{Demultiplex reads using experimental barcode}
\usage{
DemultiplexReads(fq.file, ...)
}
\arguments{
\item{fq.file}{Path to FASTQ file}

\item{...}{list of experiments and barcodes e.g. experiment_1 = "NNNN"}
}
\value{
FASTQ files as per experimental barcodes
}
\description{
Demultiplex reads using experimental barcode
}
