% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{EvaluateBarcodes}
\alias{EvaluateBarcodes}
\title{Evaluate Barcodes}
\usage{
EvaluateBarcodes(bam.dt, barcodeindex)
}
\arguments{
\item{bam.dt}{BAM data table (from LoadBAM)}

\item{barcodeindex}{Path to barcode index table}
}
\value{
data.table of unique hybrids
}
\description{
Matches arms from BAM and uses random barcode to remove PCR duplicates
}
