% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{ExportBED}
\alias{ExportBED}
\title{Export BED file}
\usage{
ExportBED(hybrid.dt, filename)
}
\arguments{
\item{hybrid.dt}{hybrid data.table}

\item{filename}{Filename to save}
}
\value{
BED file
}
\description{
Export genomic coordinate BED file from hybrid data.table
}

