% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes_mapping_functions.R
\name{ExtractHybrids}
\alias{ExtractHybrids}
\title{Remove adapters to select hybrid reads}
\usage{
ExtractHybrids(fq.file, adapterA, adapterB)
}
\arguments{
\item{fq.file}{Path to FASTQ file}

\item{adapterA}{Adapter A sequence}

\item{adapterB}{Adapter B sequence}
}
\value{
FASTQ file of L and R arms (unzipped)
}
\description{
Remove A and B adapters to select hybrid reads expanding on YS's strategy
}

