% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes_mapping_functions.R
\name{ExtractRandomBarcodes}
\alias{ExtractRandomBarcodes}
\title{Extract random barcodes and leave experiment barcode at front}
\usage{
ExtractRandomBarcodes(fq.file, barcode.filename)
}
\arguments{
\item{fq.file}{Path to FASTQ file}

\item{barcode.filename}{File for barcode index table}
}
\value{
FASTQ file and random barcode tsv
}
\description{
Extract random barcodes and leave experiment barcode at front
}
