% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnahybrid.R
\name{FindDuplex}
\alias{FindDuplex}
\title{Find Duplex starts}
\usage{
FindDuplex(hybrid.dt, RNAhybrid.out)
}
\arguments{
\item{hybrid.dt}{hybrid data.table}

\item{RNAhybrid.out}{path to RNA Hybrid output}
}
\value{
data.table of duplexes
}
\description{
Get start positions for each arm and width for longest duplex.
}
