% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnahybrid.R
\name{FindDuplexPositions}
\alias{FindDuplexPositions}
\title{Find Duplex starts}
\usage{
FindDuplexPositions(RNAhybrid)
}
\arguments{
\item{RNAhybrid}{RNA Hybrid row}
}
\value{
list of duplex starts and length
}
\description{
Get start positions for each arm and width for longest duplex. Called by FindDuplex
}

