% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{FindIslands}
\alias{FindIslands}
\title{Find confident islands}
\usage{
FindIslands(hybrid.dt, minimum.coverage, minimum.width)
}
\arguments{
\item{hybrid.dt}{hybrid data.table}

\item{minimum.coverage}{minimum coverage for an island}

\item{minimum.width}{minimum width for an island}
}
\value{
annotated hybrid data.table
}
\description{
Find confident hybrid islands
}
