% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnahybrid.R
\name{GetSequence}
\alias{GetSequence}
\title{Get sequence}
\usage{
GetSequence(hybrid.dt)
}
\arguments{
\item{hybrid.dt}{hybrid data.table}
}
\value{
hybrid data.table with sequences
}
\description{
Get sequence from annotated hybrid data.table
}

