% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{LoadBAM}
\alias{LoadBAM}
\title{Loads hybrid BAM file}
\usage{
LoadBAM(bam.file)
}
\arguments{
\item{bam.file}{Path to BAM file}
}
\value{
data.table of valid hybrids
}
\description{
Loads hybrid reads from BAM and identify valid reads (i.e both L and R arms aligned)
}
