% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PlotHybridRegion}
\alias{PlotHybridRegion}
\title{Plot hybrid source}
\usage{
PlotHybridRegion(hybrid.dt, filename)
}
\arguments{
\item{hybrid.dt}{hybrid data.table}

\item{filename}{filename for plot with extension}
}
\value{
plot and table to command line
}
\description{
Plot the hybrid region for all arms
}
