% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PlotIntraInterRatio}
\alias{PlotIntraInterRatio}
\title{Plot intra-inter ratio (for genic reads)}
\usage{
PlotIntraInterRatio(genic.dt, filename)
}
\arguments{
\item{genic.dt}{hybrid data.table of genic reads}

\item{filename}{filename for plot with extension}
}
\value{
plot
}
\description{
Plot the intra-inter ratio (for genic reads)
}

