% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PlotLoopLength}
\alias{PlotLoopLength}
\title{Plot loop length}
\usage{
PlotLoopLength(genic.dt, weighting = c(TRUE, FALSE), ratio = c(TRUE, FALSE),
  filename)
}
\arguments{
\item{genic.dt}{hybrid data.table of genic reads}

\item{weighting}{weight by coverage of islands or not}

\item{filename}{filename for plot with extension}
}
\value{
plot
}
\description{
Plot the loop length for genic reads (CDS-CDS and UTR3-UTR3)
}
