% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PlotRegionMate}
\alias{PlotRegionMate}
\title{Plot region mate}
\usage{
PlotRegionMate(hybrid.dt, region = c("miRNA", "lincRNA", "ncRNA", "CDS",
  "UTR3", "UTR5", "INTRON", "INTERGENIC", "tRNA", "rRNA", "other"), filename)
}
\arguments{
\item{hybrid.dt}{hybrid data.table}

\item{region}{region for which to find mates}

\item{filename}{filename for plot with extension}
}
\value{
plot
}
\description{
Plot the mate for a given region
}

