% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnahybrid.R
\name{RunRNAHybrid}
\alias{RunRNAHybrid}
\title{Run RNA hybrid}
\usage{
RunRNAHybrid(hybrid.dt, file)
}
\arguments{
\item{hybrid.dt}{hybrid data.table}

\item{file}{RNAhybrid output filename}
}
\value{
annotated hybrid data.table with sequences
}
\description{
Run RNA Hybrid using info from hybrid data table
}

