% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{SelectGenicHybrid}
\alias{SelectGenicHybrid}
\title{Select genic reads}
\usage{
SelectGenicHybrid(hybrid.dt, type = c("all", "intra", "inter"),
  class = c("mature", "immature"))
}
\arguments{
\item{hybrid.dt}{hybrid data.table}

\item{type}{type of genic read}

\item{class}{class of genic read (i.e. include introns or not)}
}
\value{
hybrid data.table
}
\description{
Select genic reads (only for those with both arms mapped to genes)
}
